requirements.txt
py-clob-client==0.34.6
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
sqlalchemy==2.0.36
aiohttp==3.11.0
pandas==2.2.3
numpy==2.1.3
ta==0.11.0
httpx==0.28.0
apscheduler==3.10.4
websockets==13.1
python-multipart==0.0.12
scipy==1.14.1
